feat: move the commands specific functions into their own modules
