+++ app/modules/aws/aws_account_health.py
@@ -175,3 +175,101 @@ def get_ignored_security_hub_issues():
     ]
 
     return list(map(lambda t: {"Value": t, "Comparison": "NOT_EQUALS"}, ignored_issues))
+
+
+def health_view_handler(ack, body, logger, client):
+    ack()
+
+    account_id = body["view"]["state"]["values"]["account"]["account"][
+        "selected_option"
+    ]["value"]
+
+    account_name = body["view"]["state"]["values"]["account"]["account"][
+        "selected_option"
+    ]["text"]["text"]
+
+    account_info = get_account_health(account_id)
+
+    blocks = {
+        "type": "modal",
+        "callback_id": "health_view",
+        "title": {"type": "plain_text", "text": "AWS - Health Check"},
+        "close": {"type": "plain_text", "text": "Close"},
+        "blocks": [
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": f"Health check for *{account_name}*: ({account_id})",
+                },
+            },
+            {"type": "divider"},
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": f"""
+*Cost:*
+
+{account_info['cost']['last_month']['start_date']} - {account_info['cost']['last_month']['end_date']}: ${account_info['cost']['last_month']['amount']} USD
+{account_info['cost']['current_month']['start_date']} - {account_info['cost']['current_month']['end_date']}: ${account_info['cost']['current_month']['amount']} USD
+                        """,
+                },
+            },
+            {"type": "divider"},
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": f"""
+*Security:*
+
+{"✅" if account_info['security']['config'] == 0 else "❌"} Config ({account_info['security']['config']} issues)\n
+{"✅" if account_info['security']['guardduty'] == 0 else "❌"} GuardDuty ({account_info['security']['guardduty']} issues)\n
+{"✅" if account_info['security']['securityhub'] == 0 else "❌"} SecurityHub ({account_info['security']['securityhub']} issues)\n
+                        """,
+                },
+            },
+        ],
+    }
+
+    client.views_open(
+        trigger_id=body["trigger_id"],
+        view=blocks,
+    )
+
+
+def request_health_modal(client, body):
+    accounts = get_accounts()
+    options = [
+        {
+            "text": {"type": "plain_text", "text": value},
+            "value": key,
+        }
+        for key, value in accounts.items()
+    ]
+    client.views_open(
+        trigger_id=body["trigger_id"],
+        view={
+            "type": "modal",
+            "callback_id": "aws_health_view",
+            "title": {"type": "plain_text", "text": "AWS - Account health"},
+            "submit": {"type": "plain_text", "text": "Submit"},
+            "blocks": [
+                {
+                    "block_id": "account",
+                    "type": "input",
+                    "element": {
+                        "type": "static_select",
+                        "placeholder": {
+                            "type": "plain_text",
+                            "text": "Select an account to view | Choisissez un compte à afficher",
+                        },
+                        "options": options,
+                        "action_id": "account",
+                    },
+                    "label": {"type": "plain_text", "text": "Account", "emoji": True},
+                }
+            ],
+        },
+    )
