+++ app/tests/modules/aws/test_aws_access_requests.py
@@ -1,6 +1,6 @@
 import datetime
 
-from unittest.mock import patch
+from unittest.mock import ANY, MagicMock, patch
 
 from modules.aws import aws_access_requests
 
@@ -295,3 +295,211 @@ def test_get_past_requests_empty(mock_dynamodb_scan):
             ExpressionAttributeValues={":current_time": {"S": str(current_timestamp)}},
         )
         assert items == mock_response["Items"]
+
+
+def test_access_view_handler_returns_errors_with_rational_over_2000_characters():
+    ack = MagicMock()
+    body = build_body(rationale="a" * 2001)
+
+    aws_access_requests.access_view_handler(ack, body, MagicMock(), MagicMock())
+    ack.assert_called_with(
+        response_action="errors",
+        errors={"rationale": "Please use less than 2000 characters"},
+    )
+
+
+@patch("modules.aws.aws_access_requests.log_ops_message")
+@patch("modules.aws.aws_access_requests.aws_sso.get_user_id")
+def test_access_view_handler_with_unknown_sso_user(
+    get_user_id_mock, log_ops_message_mock
+):
+    ack = MagicMock()
+    body = build_body()
+    logger = MagicMock()
+    client = MagicMock()
+
+    client.users_info.return_value = {
+        "user": {"id": "user_id", "profile": {"email": "email"}}
+    }
+
+    get_user_id_mock.return_value = None
+
+    aws_access_requests.access_view_handler(ack, body, logger, client)
+    ack.assert_called()
+    client.users_info.assert_called_with(user="user_id")
+    logger.info.assert_called_with(
+        "<@user_id> (email) requested access to account (account_id) with access_type priviliges.\n\nRationale: rationale"
+    )
+    log_ops_message_mock.assert_called_with(
+        client,
+        "<@user_id> (email) requested access to account (account_id) with access_type priviliges.\n\nRationale: rationale",
+    )
+    client.chat_postEphemeral.assert_called_with(
+        channel="user_id",
+        user="user_id",
+        text="<@user_id> (email) is not registered with AWS SSO. Please contact your administrator.\n<@user_id> (email) n'est pas enregistré avec AWS SSO. SVP contactez votre administrateur.",
+    )
+
+
+@patch("modules.aws.aws_access_requests.log_ops_message")
+@patch("modules.aws.aws_access_requests.aws_sso.get_user_id")
+@patch("modules.aws.aws_access_requests.already_has_access")
+def test_access_view_handler_with_existing_access(
+    already_has_access_mock, get_user_id_mock, log_ops_message_mock
+):
+    ack = MagicMock()
+    body = build_body()
+    logger = MagicMock()
+    client = MagicMock()
+
+    client.users_info.return_value = {
+        "user": {"id": "user_id", "profile": {"email": "email"}}
+    }
+
+    get_user_id_mock.return_value = "aws_user_id"
+    already_has_access_mock.return_value = 10
+
+    aws_access_requests.access_view_handler(ack, body, logger, client)
+    ack.assert_called()
+    client.users_info.assert_called_with(user="user_id")
+    logger.info.assert_called_with(
+        "<@user_id> (email) requested access to account (account_id) with access_type priviliges.\n\nRationale: rationale"
+    )
+    log_ops_message_mock.assert_called_with(
+        client,
+        "<@user_id> (email) requested access to account (account_id) with access_type priviliges.\n\nRationale: rationale",
+    )
+    client.chat_postEphemeral.assert_called_with(
+        channel="user_id",
+        user="user_id",
+        text="You already have access to account (account_id) with access type access_type. Your access will expire in 10 minutes.",
+    )
+
+
+@patch("modules.aws.aws_access_requests.log_ops_message")
+@patch("modules.aws.aws_access_requests.aws_sso.get_user_id")
+@patch("modules.aws.aws_access_requests.already_has_access")
+@patch("modules.aws.aws_access_requests.create_aws_access_request")
+@patch("modules.aws.aws_access_requests.aws_sso.add_permissions_for_user")
+def test_access_view_handler_successful_access_request(
+    add_permissions_for_user_mock,
+    create_aws_access_request_mock,
+    already_has_access_mock,
+    get_user_id_mock,
+    log_ops_message_mock,
+):
+    ack = MagicMock()
+    body = build_body()
+    logger = MagicMock()
+    client = MagicMock()
+
+    client.users_info.return_value = {
+        "user": {"id": "user_id", "profile": {"email": "email"}}
+    }
+
+    get_user_id_mock.return_value = "aws_user_id"
+    already_has_access_mock.return_value = False
+    create_aws_access_request_mock.return_value = True
+    add_permissions_for_user_mock.return_value = True
+
+    aws_access_requests.access_view_handler(ack, body, logger, client)
+    ack.assert_called()
+    client.users_info.assert_called_with(user="user_id")
+    logger.info.assert_called_with(
+        "<@user_id> (email) requested access to account (account_id) with access_type priviliges.\n\nRationale: rationale"
+    )
+    log_ops_message_mock.assert_called_with(
+        client,
+        "<@user_id> (email) requested access to account (account_id) with access_type priviliges.\n\nRationale: rationale",
+    )
+    client.chat_postEphemeral.assert_called_with(
+        channel="user_id",
+        user="user_id",
+        text="Provisioning access_type access request for account (account_id). This can take a minute or two. Visit <https://cds-snc.awsapps.com/start#/|https://cds-snc.awsapps.com/start#/> to gain access.\nTraitement de la requête d'accès access_type pour le compte account (account_id) en cours. Cela peut prendre quelques minutes. Visitez <https://cds-snc.awsapps.com/start#/|https://cds-snc.awsapps.com/start#/> pour y accéder",
+    )
+
+
+@patch("modules.aws.aws_access_requests.log_ops_message")
+@patch("modules.aws.aws_access_requests.aws_sso.get_user_id")
+@patch("modules.aws.aws_access_requests.already_has_access")
+@patch("modules.aws.aws_access_requests.create_aws_access_request")
+@patch("modules.aws.aws_access_requests.aws_sso.add_permissions_for_user")
+def test_access_view_handler_failed_access_request(
+    add_permissions_for_user_mock,
+    create_aws_access_request_mock,
+    already_has_access_mock,
+    get_user_id_mock,
+    log_ops_message_mock,
+):
+    ack = MagicMock()
+    body = build_body()
+    logger = MagicMock()
+    client = MagicMock()
+
+    client.users_info.return_value = {
+        "user": {"id": "user_id", "profile": {"email": "email"}}
+    }
+
+    get_user_id_mock.return_value = "aws_user_id"
+    already_has_access_mock.return_value = False
+    create_aws_access_request_mock.return_value = True
+    add_permissions_for_user_mock.return_value = False
+
+    aws_access_requests.access_view_handler(ack, body, logger, client)
+    ack.assert_called()
+    client.users_info.assert_called_with(user="user_id")
+    logger.info.assert_called_with(
+        "<@user_id> (email) requested access to account (account_id) with access_type priviliges.\n\nRationale: rationale"
+    )
+    log_ops_message_mock.assert_called_with(
+        client,
+        "<@user_id> (email) requested access to account (account_id) with access_type priviliges.\n\nRationale: rationale",
+    )
+    client.chat_postEphemeral.assert_called_with(
+        channel="user_id",
+        user="user_id",
+        text="Failed to provision access_type access request for account (account_id). Please drop a note in the <#sre-and-tech-ops> channel.\nLa requête d'accès access_type pour account (account_id) a échouée. Envoyez une note sur le canal <#sre-and-tech-ops>",
+    )
+
+
+@patch("modules.aws.aws.aws_sso.get_accounts")
+def test_request_access_modal(get_accounts_mock):
+    get_accounts_mock.return_value = {"id": "name"}
+
+    client = MagicMock()
+    body = {"trigger_id": "trigger_id", "view": {"state": {"values": {}}}}
+
+    aws_access_requests.request_access_modal(client, body)
+    client.views_open.assert_called_with(
+        trigger_id="trigger_id",
+        view=ANY,
+    )
+
+
+def build_body(
+    rationale="rationale",
+    account="account",
+    account_id="account_id",
+    access_type="access_type",
+):
+    return {
+        "view": {
+            "state": {
+                "values": {
+                    "rationale": {"rationale": {"value": rationale}},
+                    "account": {
+                        "account": {
+                            "selected_option": {
+                                "text": {"text": account},
+                                "value": account_id,
+                            }
+                        }
+                    },
+                    "access_type": {
+                        "access_type": {"selected_option": {"value": access_type}}
+                    },
+                }
+            }
+        },
+        "user": {"id": "user_id"},
+    }
